nidsmonday/monday/monday-learning-amalgamated/python-services/main.py
from fastapi import FastAPI, HTTPException
from pydantic import BaseModel
from typing import List, Optional, Dict, Any
import os
from dotenv import load_dotenv

load_dotenv()

app = FastAPI(
    title="Monday Python Services",
    description="Research synthesis and ML processing services for Monday: Learning, Amalgamated",
    version="1.0.0"
)

# Data models
class ResearchRequest(BaseModel):
    query: str
    sources: List[Dict[str, Any]]
    mode: str = "synthesis"

class KnowledgeGraphRequest(BaseModel):
    concepts: List[str]
    relationships: List[Dict[str, str]]

class VisualizationRequest(BaseModel):
    concept: str
    type: str = "auto"
    complexity: int = 1

# Health check endpoint
@app.get("/health")
async def health_check():
    return {
        "status": "healthy",
        "service": "monday-python-services",
        "version": "1.0.0"
    }

# Research synthesis endpoint
@app.post("/research/synthesize")
async def synthesize_research(request: ResearchRequest):
    """
    Synthesize multiple research sources into a unified knowledge representation
    """
    try:
        # Placeholder for research synthesis logic
        synthesis_result = {
            "query": request.query,
            "synthesis": f"Synthesized research for: {request.query}",
            "key_concepts": ["concept1", "concept2", "concept3"],
            "connections": [],
            "confidence": 0.85,
            "sources_processed": len(request.sources)
        }
        
        return synthesis_result
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

# Knowledge graph processing
@app.post("/knowledge/graph")
async def process_knowledge_graph(request: KnowledgeGraphRequest):
    """
    Process concepts and relationships into a knowledge graph structure
    """
    try:
        # Placeholder for knowledge graph processing
        graph_result = {
            "nodes": [{"id": concept, "type": "concept"} for concept in request.concepts],
            "edges": request.relationships,
            "layout": "force-directed",
            "metadata": {
                "node_count": len(request.concepts),
                "edge_count": len(request.relationships)
            }
        }
        
        return graph_result
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

# Visualization generation
@app.post("/visualization/generate")
async def generate_visualization(request: VisualizationRequest):
    """
    Generate 3D visualization data for concepts
    """
    try:
        # Placeholder for visualization generation
        viz_result = {
            "concept": request.concept,
            "type": request.type,
            "geometry": {
                "vertices": [],
                "faces": [],
                "normals": []
            },
            "materials": {
                "diffuse": "#20808D",
                "specular": "#FBFAF4",
                "opacity": 0.8
            },
            "animations": {
                "intro": "fade_in",
                "idle": "gentle_rotate",
                "exit": "fade_out"
            },
            "complexity_level": request.complexity
        }
        
        return viz_result
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

# Concept analysis endpoint
@app.post("/analysis/concept")
async def analyze_concept(concept: str):
    """
    Analyze a concept for complexity, prerequisites, and related topics
    """
    try:
        analysis_result = {
            "concept": concept,
            "complexity_score": 0.6,
            "prerequisites": [],
            "related_concepts": [],
            "learning_difficulty": "intermediate",
            "estimated_time": "15 minutes",
            "visualization_type": "tree"
        }
        
        return analysis_result
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

# Learning path optimization
@app.post("/learning/optimize-path")
async def optimize_learning_path(concepts: List[str], user_level: str = "beginner"):
    """
    Optimize the learning path for a set of concepts based on user level
    """
    try:
        optimized_path = {
            "path": concepts,  # Placeholder - would be reordered based on difficulty
            "user_level": user_level,
            "estimated_total_time": len(concepts) * 10,
            "difficulty_progression": "gradual",
            "checkpoints": [len(concepts) // 2] if len(concepts) > 2 else [],
            "recommendations": [
                "Start with basic concepts",
                "Take breaks every 20 minutes",
                "Practice active recall"
            ]
        }
        
        return optimized_path
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000) 