nidsmonday/monday/monday-learning-amalgamated/python-services/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
nltk==3.8.1
transformers==4.35.2
torch==2.1.1
redis==5.0.1
psycopg2-binary==2.9.9
neo4j==5.14.1
requests==2.31.0
aiohttp==3.9.1
celery==5.3.4
python-multipart==0.0.6
python-dotenv==1.0.0 